Add files via upload
This Calculator was a result of having watched the Codemy video called, 
'Build A Simple Calculator App - Python Tkinter GUI Tutorial #5' on youtube.  
I added the rest of the button functionality, and the History / Memory functionality as well.

At this point, the calculator works, but is not able to chain calculations without
hitting the equal button between them. In other words, you can't do 2*2*2 = 8, you 
must do 2*2 = 4*2 = 8. I have not figured that part out... yet. 

You can use the History button to view the last 8 equations calculated. For each click
of the History button, you see another equation among the last x number of equations.  
If you click the History button more than 8 times, or more than the number of equations 
calculated, you get an error message.

By the way, I realize this is not the most optimal or DRY way to build an app.  
There's tons of repeated code, most of which was cut and pasted. If I had found a way to
put each equation scenario into a function, with changing the names of the equations by 
just their numbers:  equation1, then equation2 etc, that might have helped. 
(I tried using a list, and then dictionary, but I ran into problems with both approaches)